Add files via upload
This project displays the simulation of Apple Inc. Stock price to help explain the impact of risk and uncertainty in prediction and forecasting the model. Since, Monte Carlo simulation takes the variable that has uncertainty and assigns it a random value, the uncertain variable in this scenario is the stock price that is volatile. Then the simulation model is run, and a result is provided. This process is repeated several times by assigning different random values to the uncertain variable( Stock Price).Once the simulation is complete, the results are averaged to arrive at an estimate.