Add files via upload
The Project focuses on creating a model that uses the price of Amazon (AMZN) stocks over the 10 year time period from 2013 to 2023 to predict the future expected price of Amazon Stocks in the USA Market. The model uses different machine learning tools to prepare the data, build model and communicate the results for decision making by investors.